chore(shared): move remootio client to shared module
