+++ RemootioAndroidAuto/automotive/build.gradle
@@ -14,7 +14,7 @@ android {
     defaultConfig {
         applicationId "chamilton0.remootioandroidws"
         minSdk 28
-        targetSdk 33
+        targetSdk 34
         versionCode 5
         versionName "0.2.0"
 
@@ -56,10 +56,10 @@ dependencies {
     implementation "androidx.car.app:app:1.2.0"
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'androidx.constraintlayout:constraintlayout-core:1.0.4'
-    implementation 'androidx.lifecycle:lifecycle-common-java8:2.6.1'
+    implementation 'androidx.lifecycle:lifecycle-common-java8:2.7.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'org.java-websocket:Java-WebSocket:1.5.3'
-    implementation 'androidx.core:core-ktx:1.10.1'
+    implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'org.apache.commons:commons-text:1.10.0'
     implementation project(path: ':shared')
     testImplementation 'junit:junit:4.13.2'
