+++ RemootioAndroidAuto/main/build.gradle
@@ -10,7 +10,7 @@ android {
     defaultConfig {
         applicationId "chamilton0.remootioandroidws"
         minSdk 28
-        targetSdk 33
+        targetSdk 34
         versionCode 1
         versionName "1.0"
