+++ RemootioAndroidAuto/shared/src/main/java/chamilton0/remootioandroidws/shared/RemootioClient.kt
@@ -1,4 +1,4 @@
-package chamilton0.remootioandroidws.auto
+package chamilton0.remootioandroidws.shared
 
 import java.net.URI
 import org.java_websocket.client.WebSocketClient
